refactor: height type을 number -> string으로 수정하여 타입 통일화
