+++ apps/client/components/ui/types/commonElements.ts
@@ -12,3 +12,26 @@ export type Text = {
   fontWeight?: string;
   margin?: string;
 };
+
+export type Wrap = {
+  width?: string;
+  height?: string;
+  margin?: string;
+  overflow?: string;
+  display?: string;
+  flexWrap?: string;
+  position?: string;
+  right?: string;
+  bottom?: string;
+  padding?: string;
+  flexDirection?: string;
+  justifyContent?: string;
+  alignItems?: string;
+  backgroundColor?: string;
+  borderRadius?: string;
+  show?: {
+    setShowSideMenu: (arg0: boolean) => void;
+    showSideMenu: boolean;
+  };
+  onClick?: () => void;
+};
